test(e2e): Drop duplicate nuxt-sourcemaps app in favour of nuxt-4-sourcemaps
The nuxt-4-sourcemaps app merged in #23518 covers the same ground with more
variants, so keep that one and port over the assertions it did not have:
the release create body, the release finalize call, the chunk-upload options
request, and the release reference plus checksum shape on the assemble body.
