--- dev-packages/e2e-tests/test-applications/nuxt-sourcemaps/.gitignore
@@ -1,11 +1,0 @@
-.output
-.data
-.nuxt
-.nitro
-.cache
-dist
-node_modules
-.tmp_mock_uploads.json
-.tmp_chunks
-.tmp_build_stdout
-.tmp_build_stderr
