--- dev-packages/e2e-tests/test-applications/nuxt-sourcemaps/app/app.vue
@@ -1,7 +1,0 @@
-<template>
-  <h1>{{ message }}</h1>
-</template>
-
-<script setup lang="ts">
-const message = `hello from nuxt sourcemaps test at ${new Date().toISOString()}`;
-</script>
