--- dev-packages/e2e-tests/test-applications/nuxt-sourcemaps/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "nuxt-sourcemaps",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "node start-mock-sentry-server.mjs & nuxt build > .tmp_build_stdout 2> .tmp_build_stderr; BUILD_EXIT=$?; kill %1 2>/dev/null; exit $BUILD_EXIT",
-    "clean": "npx nuxi cleanup",
-    "test:build": "pnpm install && pnpm build",
-    "test:assert": "pnpm tsx assert-build.ts"
-  },
-  "dependencies": {
-    "@sentry/nuxt": "file:../../packed/sentry-nuxt-packed.tgz",
-    "nuxt": "^4.1.2"
-  },
-  "devDependencies": {
-    "@playwright/test": "~1.56.0",
-    "@sentry-internal/test-utils": "link:../../../test-utils",
-    "tsx": "^4.23.0",
-    "typescript": "^5.6.3"
-  },
-  "volta": {
-    "extends": "../../package.json",
-    "node": "22.20.0"
-  }
-}
