--- dev-packages/e2e-tests/test-applications/nuxt-sourcemaps/sentry.client.config.ts
@@ -1,6 +1,0 @@
-import * as Sentry from '@sentry/nuxt';
-import { useRuntimeConfig } from '#imports';
-
-Sentry.init({
-  dsn: useRuntimeConfig().public.sentry.dsn,
-});
