--- dev-packages/e2e-tests/test-applications/nuxt-sourcemaps/sentry.server.config.ts
@@ -1,5 +1,0 @@
-import * as Sentry from '@sentry/nuxt';
-
-Sentry.init({
-  dsn: 'https://public@dsn.ingest.sentry.io/1337',
-});
