--- dev-packages/e2e-tests/test-applications/nuxt-sourcemaps/start-mock-sentry-server.mjs
@@ -1,3 +1,0 @@
-import { startMockSentryServer } from '@sentry-internal/test-utils';
-
-startMockSentryServer();
