--- dev-packages/e2e-tests/test-applications/nuxt-sourcemaps/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  // https://nuxt.com/docs/guide/concepts/typescript
-  "extends": "./.nuxt/tsconfig.json"
-}
